+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - For the `parse` command, added the `--profiler` and `--bench` options
   to help debugging performance issues.
+- Support for the `do` command in the jinja templater.
 - Proper parsing of the concatenate operator (`||`).
 - Proper indent handling of closing brackets.
 - Logging and benchmarking of parse performance as part of the CI pipeline.